+++ cluster/anchnet/addons/addons-start.sh
@@ -20,8 +20,13 @@
 # It is assumed that required addons yaml files are copied to ~/kube
 # directory.
 
+set -o errexit
+set -o nounset
+set -o pipefail
+
 ENABLE_CLUSTER_DNS=${ENABLE_CLUSTER_DNS:-false}
 ENABLE_CLUSTER_LOGGING=${ENABLE_CLUSTER_LOGGING:-false}
+ENABLE_CLUSTER_UI=${ENABLE_CLUSTER_UI:-false}
 SYSTEM_NAMESPACE=${SYSTEM_NAMESPACE:-"kube-system"}
 
 # Do retries when failing in objects creation from yaml files. It
@@ -61,6 +66,9 @@ function prepare-addons {
   mkdir -p ~/kube/addons/logging
   mv ~/kube/elasticsearch-rc.yaml ~/kube/elasticsearch-svc.yaml ~/kube/addons/logging
   mv ~/kube/kibana-rc.yaml ~/kube/kibana-svc.yaml ~/kube/addons/logging
+  # Kube-ui addon.
+  mkdir -p ~/kube/addons/kube-ui
+  mv ~/kube/kube-ui-rc.yaml ~/kube/kube-ui-svc.yaml ~/kube/addons/kube-ui
 }
 
 function create-dns-addon {
@@ -75,6 +83,12 @@ function create-logging-addon {
   done
 }
 
+function create-kube-ui-addon {
+  for obj in $(find ~/kube/addons/kube-ui -type f -name \*.yaml -o -name \*.json); do
+    create-resource-from-file ${obj} 10 10 "${SYSTEM_NAMESPACE}"
+  done
+}
+
 
 prepare-addons
 
@@ -84,4 +98,8 @@ fi
 
 if [[ "${ENABLE_CLUSTER_LOGGING}" == "true" ]]; then
   create-logging-addon
+fi
+
+if [[ "${ENABLE_CLUSTER_UI}" == "true" ]]; then
+  create-kube-ui-addon
 fi